Finished refactor, reenabled CA1034
