+++ Source/Libraries/CorruptCore/Corruption Engines/RTC_CustomEngine.cs
@@ -8,6 +8,7 @@
     using System.Numerics;
     using System.Windows.Forms;
     using RTCV.NetCore;
+    using RTCV.NetCore.NetCore_Extensions;
 
     public static class RTC_CustomEngine
     {
