+++ Source/Libraries/NetCore/NetCore.csproj
@@ -64,7 +64,6 @@
     <Compile Include="AllSpec.cs" />
     <Compile Include="Connector.cs" />
     <Compile Include="NetCore_Extensions.cs" />
-    <Compile Include="NetCore_Extensions2.cs" />
     <Compile Include="DebugInfo\CloudDebug_Form.cs">
       <SubType>Form</SubType>
     </Compile>
